Extract row → model helpers into `row_mapping.py`
`routes.py` used to carry 250 lines of pure row-shaping after its
endpoints. Move them into a dedicated `row_mapping` module so `routes.py`
reads as a list of endpoints, and the row-shaping is one file the two
concerns don't sit on top of each other in.

Also alias `fastapi.Path` to `PathParam` so it doesn't shadow the
`pathlib.Path` import that appears above it.
